added post show

+++ app/Http/Controllers/PostController.php
@@ -9,45 +9,69 @@ class PostController extends Controller
 {
     public function index( Request $request )
     {
+        if ( $request->wantsJson() ) {
             $post = Post::all();
 
             return response()->json($post, 200);
+        }
+
+        return view('welcome');
     }
 
     public function store( Request $request )
     {
+        if ( $request->wantsJson() ) {
+        }
+
+        return view('welcome');
         Post::create($request->all());
 
         return response()->json('Created', 200);
     }
 
-    public function show( $id )
+    public function show( $id, Request $request )
     {
-        $post = Post::find($id);
+        if ( $request->wantsJson() ) {
+            $post = Post::find($id);
 
-        return response()->json($post, 200);
+            return response()->json($post, 200);
+        }
+
+        return view('welcome');
     }
 
-    public function edit( $id )
+    public function edit( $id, Request $request )
     {
-        $post = Post::find($id);
+        if ( $request->wantsJson() ) {
+            $post = Post::find($id);
 
-        return response()->json($post, 200);
+            return response()->json($post, 200);
+        }
+
+        return view('welcome');
     }
 
     public function update( $id, Request $request )
     {
-        $post = Post::find($id);
-        $post->update($request->all());
+        if ( $request->wantsJson() ) {
+            $post = Post::find($id);
+            $post->update($request->all());
 
-        return response()->json('Updated', 200);
+            return response()->json('Updated', 200);
+        }
+
+        return view('welcome');
     }
 
-    public function destroy( $id )
+    public function destroy( $id, Request $request )
     {
-        $post = Post::find($id);
-        $post->delete();
+        if ( $request->wantsJson() ) {
+            $post = Post::find($id);
+            $post->delete();
 
-        return response()->json('Deleted', 200);
+            return response()->json('Deleted', 200);
+        }
+
+        return view('welcome');
     }
 }
